refactor(ci): extract packed SSR consumer fixtures
Move the standalone consumer project, browser probes, and page fragments
into readable fixture files while keeping dependency coordinates and
build ids in the driver.

Route fixture changes to the packed-consumer CI scope and exclude the
consumer-only sources from workspace TypeScript and Knip analysis.
