+++ scripts/fixtures/packed-ssr-consumer/index.html
@@ -1,0 +1,164 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <title>Packed consumer</title>
+    <script>
+      window.__adoptedFrameDocuments = []
+      window.__recordAdoptedFrame = function (frameDocument) {
+        window.__adoptedFrameDocuments.push(frameDocument)
+      }
+      window.__parserOwnedConnectionChildCounts = []
+      window.__parserChildDisconnections = 0
+      class ParserChild extends HTMLElement {
+        connectedCallback() {
+          this.ownerHost = this.parentElement
+        }
+        disconnectedCallback() {
+          window.__parserChildDisconnections++
+          var ownerHost = this.ownerHost
+          if (ownerHost === null || ownerHost === undefined) {
+            return
+          }
+          ownerHost.removeAttribute('title')
+          var reinserted = document.createElement('span')
+          reinserted.setAttribute('data-parser-reinserted', '')
+          ownerHost.appendChild(reinserted)
+        }
+      }
+      customElements.define('x-parser-child', ParserChild)
+      class ParserOwned extends HTMLElement {
+        connectedCallback() {
+          this.ownerParent = this.parentElement
+          this.ownerEarlierSibling = this.previousElementSibling
+          window.__parserOwnedConnectionChildCounts.push(this.childNodes.length)
+          if (this.childNodes.length > 0) {
+            return
+          }
+          var child = document.createElement('x-parser-child')
+          child.id = 'parser-view-child'
+          child.textContent = 'view'
+          this.componentOwnedChild = child
+          this.appendChild(child)
+        }
+        disconnectedCallback() {
+          this.ownerParent?.removeAttribute('title')
+          this.ownerEarlierSibling?.removeAttribute('title')
+          var earlierText = this.ownerEarlierSibling?.firstChild
+          if (earlierText !== null && earlierText !== undefined) {
+            earlierText.data = 'component-mutated'
+          }
+        }
+        mutateOwned() {
+          this.componentOwnedChild.textContent = 'component-mutated'
+        }
+      }
+      customElements.define('x-parser-owned', ParserOwned)
+    </script>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script>
+      window.__probe = (function () {
+        var root = document.querySelector('[data-foldkit-app]')
+        var field = document.querySelector('#field')
+        if (field !== null) {
+          field.value = '{{TYPED_VALUE}}'
+        }
+        window.__observedIdChanges = []
+        window.__customElementConstructions = { holder: 0, inner: 0 }
+        window.__idPropertyWrites = 0
+        window.__innerHtmlPropertyWrites = 0
+        class ObservedId extends HTMLElement {
+          constructor() {
+            super()
+            window.__customElementConstructions.holder++
+          }
+          static get observedAttributes() {
+            return ['id', 'dir']
+          }
+          get id() {
+            return 'component-id'
+          }
+          set id(value) {
+            window.__idPropertyWrites++
+          }
+          get innerHTML() {
+            return 'component markup'
+          }
+          set innerHTML(value) {
+            window.__innerHtmlPropertyWrites++
+          }
+          attributeChangedCallback(name, oldValue, newValue) {
+            window.__observedIdChanges.push([name, oldValue, newValue])
+          }
+        }
+        customElements.define('x-observed-id', ObservedId)
+        class InnerProbe extends HTMLElement {
+          constructor() {
+            super()
+            window.__customElementConstructions.inner++
+          }
+        }
+        customElements.define('x-inner-probe', InnerProbe)
+        var styled = document.querySelector('#styled')
+        window.__styleMutations = []
+        var styleObserver = new MutationObserver(function (records) {
+          window.__styleMutations.push.apply(window.__styleMutations, records)
+        })
+        window.__readStyleMutationCount = function () {
+          window.__styleMutations.push.apply(
+            window.__styleMutations,
+            styleObserver.takeRecords(),
+          )
+          return window.__styleMutations.length
+        }
+        if (styled !== null) {
+          styleObserver.observe(styled, {
+            attributes: true,
+            attributeFilter: ['style'],
+          })
+        }
+        var controlledNodes = {}
+        for (var id of [
+          'equal-value',
+          'equal-checked',
+          'raw-select',
+          'released-textarea',
+          'released-output',
+          'inner-select',
+          'released-file',
+          'released-size',
+          'released-tabindex',
+          'released-dimensions',
+          'released-start',
+          'styled',
+          'observed-id',
+          'custom-inner-probe',
+          'native-inner-html',
+          'native-inner-probe',
+        ]) {
+          controlledNodes[id] = document.getElementById(id)
+        }
+        var parserOwnedHost = document.getElementById('parser-owned')
+        return {
+          root: root,
+          field: field,
+          controlledNodes: controlledNodes,
+          observedIdChangesAtDefinition: window.__observedIdChanges.length,
+          customElementConstructionsAtDefinition: {
+            holder: window.__customElementConstructions.holder,
+            inner: window.__customElementConstructions.inner,
+          },
+          adoptedFrame: document.getElementById('adopted-frame'),
+          parserOwnedHost: parserOwnedHost,
+          parserComponentChild: parserOwnedHost.componentOwnedChild,
+          parserServerChild: parserOwnedHost.lastElementChild,
+          parserChildrenBeforeHydration: parserOwnedHost.childElementCount,
+          styleObserver: styleObserver,
+        }
+      })()
+    </script>
+    <script type="module" src="/src/entry.ts"></script>
+  </body>
+</html>
