+++ scripts/fixtures/packed-ssr-consumer/pages/lifecycle-body.html
@@ -1,0 +1,2 @@
+<iframe id="probe-frame" src="{{FRAME_PATH}}"></iframe>
+<x-reconnect-probe></x-reconnect-probe>
