+++ scripts/fixtures/packed-ssr-consumer/pages/modal-body.html
@@ -1,0 +1,7 @@
+<iframe id="modal-frame" src="{{FRAME_PATH}}"></iframe>
+<dialog id="light-modal" data-foldkit-refusal-shield data-foldkit-refused>
+  <input />
+  <a href="/modal-navigated">Navigate</a>
+</dialog>
+<div id="open-modal-host"></div>
+<x-closed-modal-host></x-closed-modal-host>
