+++ scripts/fixtures/packed-ssr-consumer/src/vite-env.d.ts
@@ -1,0 +1,9 @@
+/// <reference types="vite/client" />
+
+interface ImportMetaEnv {
+  readonly FOLDKIT_BUILD_ID?: string
+}
+
+interface ImportMeta {
+  readonly env: ImportMetaEnv
+}
