+++ scripts/fixtures/packed-ssr-consumer/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "compilerOptions": {
+    "strict": true,
+    "noEmit": true,
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
+    "target": "es2022",
+    "lib": ["esnext", "dom"],
+    "types": []
+  },
+  "include": ["src/packed-types.ts"]
+}
