+++ scripts/fixtures/packed-ssr-consumer/vite.config.ts
@@ -1,0 +1,7 @@
+import { defineConfig } from 'vite'
+
+import { foldkit } from '@foldkit/vite-plugin'
+
+export default defineConfig({
+  plugins: [foldkit({ ssr: { serverEntry: '/src/entry.server.ts' } })],
+})
