+++ scripts/tsconfig.json
@@ -6,5 +6,6 @@
     "noEmit": true,
     "types": ["node"]
   },
-  "include": ["*.ts", "*.mts", "fixtures/**/*.ts"]
+  "include": ["*.ts", "*.mts", "fixtures/**/*.ts"],
+  "exclude": ["fixtures/packed-ssr-consumer"]
 }
